Updated mongodb, removed mongodb-promise, added .tscparams
