+++ mongodb/mongodb.d.ts.tscparams
@@ -1,0 +1,1 @@
+--target ES6
